line_count: refactor code
